+++ src/textual/app.py
@@ -12,7 +12,12 @@ import warnings
 import signal
 from asyncio import Task
 from concurrent.futures import Future
-from contextlib import asynccontextmanager, redirect_stderr, redirect_stdout
+from contextlib import (
+    asynccontextmanager,
+    contextmanager,
+    redirect_stderr,
+    redirect_stdout,
+)
 from datetime import datetime
 from functools import partial
 from pathlib import Path, PurePath
@@ -25,6 +30,7 @@ from typing import (
     Callable,
     Generic,
     Iterable,
+    Iterator,
     List,
     Type,
     TypeVar,
@@ -157,6 +163,10 @@ class _WriterThread(threading.Thread):
         self._queue: Queue[str | None] = Queue(MAX_QUEUED_WRITES)
         self._file = sys.__stdout__
 
+        self._flushed = threading.Event()
+        self._resume = threading.Event()
+        self._resume.set()
+
     def write(self, text: str) -> None:
         """Write text. Text will be enqueued for writing.
 
@@ -201,6 +211,16 @@ class _WriterThread(threading.Thread):
             write(text)
             if empty:
                 flush()
+                self._flushed.set()
+                self._resume.wait()
+                self._flushed.clear()
+
+    @contextmanager
+    def paused(self) -> Iterator[None]:
+        self._resume.clear()
+        self._flushed.wait()
+        yield
+        self._resume.set()
 
     def stop(self) -> None:
         """Stop the thread, and block until it finished."""
@@ -980,6 +1000,28 @@ class App(Generic[ReturnType], DOMNode):
             event_loop = asyncio.get_event_loop()
             event_loop.run_until_complete(run_app())
         return self.return_value
+
+    @contextmanager
+    def suspend(self) -> Iterator[None]:
+        """
+        A context manager that temporarily suspends the app while inside the managed block.
+
+        While inside the `with` block, the app will stop reading input and emitting output.
+        Other applications will have full control of the terminal,
+        configured as it was before the app started running.
+
+        When the `with` block ends, the application will start reading input and emitting output again.
+        """
+        driver = self._driver
+        if driver is not None:
+            driver.stop_application_mode()
+
+            with self._writer_thread.paused(), redirect_stdout(
+                sys.__stdout__
+            ), redirect_stderr(sys.__stderr__):
+                yield
+
+            driver.start_application_mode()
 
     async def _on_css_change(self) -> None:
         """Called when the CSS changes (if watch_css is True)."""