+++ src/textual/drivers/win32.py
@@ -157,8 +157,8 @@ def enable_application_mode() -> Callable[[], None]:
         A callable that will restore terminal to previous state.
     """
 
-    terminal_in = sys.stdin
-    terminal_out = sys.stdout
+    terminal_in = sys.__stdin__
+    terminal_out = sys.__stdout__
 
     current_console_mode_in = _get_console_mode(terminal_in)
     current_console_mode_out = _get_console_mode(terminal_out)